fix: use client provided uuid for `File.id` during object creation
`FileForm` ignores explicitly provided value of field `id` because
of `editable=False` argument in model definition. So, when a new
file object is created using `file_form.save()` a randomly new
value is generated for `id`. This results in key mismatch
with the file stored in the object storage.

Using `FileForm` to validate the data and then creating the file
object using `File.objects.create()` passing in the id explicitly
fixes the bug.