fix: unauthorized file and directory operations
Add ownership checks for delete and edit operations.
When file/directory is accessed by a user other than its owner, rename and
delete options are not visible on the client. Its implementation is very
basic and does not yet work with directories with two 2 or more levels of
hierarchy. Although the operations are still correctly validated on the
server side.